tests: move matcher tests to test_matchers
